@@ -55,80 +66,82 @@
   },
   "homepage": "https://github.com/steveblue/custom-elements#readme",
   "dependencies": {
-    "@skatejs/ssr": "^0.19.11",
-    "@ungap/custom-elements": "^0.1.8",
-    "broadcastchannel-polyfill": "^1.0.0",
-    "chalk": "^3.0.0",
+    "@lit-labs/ssr": "^3.2.2",
+    "@ungap/custom-elements": "^1.3.0",
+    "broadcastchannel-polyfill": "^1.0.1",
+    "chalk": "^5.3.0",
+    "cheerio": "^1.0.0",
     "compression": "^1.7.4",
-    "concurrently": "^5.1.0",
+    "concurrently": "^9.0.1",
     "cors": "^2.8.5",
-    "element-internals-polyfill": "^0.1.9",
-    "express": "^4.17.1",
-    "html-minifier-terser": "^5.1.1",
+    "cross-env": "^7.0.3",
+    "element-internals-polyfill": "^1.3.12",
+    "express": "^4.21.0",
+    "gh-pages": "^6.1.1",
+    "he": "^1.2.0",
+    "html-minifier-terser": "^7.2.0",
     "http": "^0.0.0",
     "https": "^1.0.0",
-    "parcel-bundler": "^1.12.4",
-    "parcel-plugin-css-as-string": "^1.0.1",
-    "parcel-plugin-externals": "^0.5.1",
-    "prismjs": "^1.27.0",
-    "rimraf": "^3.0.2"
+    "jsdom": "^25.0.1",
+    "lit-html": "^3.2.0",
+    "node-fetch": "2.6.7",
+    "node-window-polyfill": "^1.0.2",
+    "prismjs": "^1.29.0",
+    "rimraf": "^6.0.1",
+    "vite-node": "^2.1.2"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^11.1.0",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^7.1.3",
-    "@rollup/plugin-replace": "^2.3.3",
-    "@rollup/plugin-typescript": "^4.1.1",
-    "@types/node": "^13.7.4",
-    "@typescript-eslint/eslint-plugin": "^2.19.2",
-    "@typescript-eslint/parser": "^2.19.2",
-    "autoprefixer": "^9.7.4",
-    "chokidar": "^3.4.0",
-    "chromedriver": "^101.0.0",
-    "copyfiles": "^2.4.0",
-    "cssnano": "^4.1.10",
-    "cypress": "^5.3.0",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.10.0",
-    "eslint-plugin-prettier": "^3.1.2",
-    "gh-pages": "^3.1.0",
-    "htmlnano": "^0.2.7",
-    "http-server": "^0.12.3",
-    "husky": "^4.3.0",
-    "lint-staged": "^10.5.0",
-    "node-sass": "^7.0.0",
-    "posthtml": "^0.13.4",
-    "posthtml-inline": "^1.0.0",
-    "prettier": "^1.19.1",
-    "rollup": "2.3.2",
+    "@eslint/js": "^9.12.0",
+    "@rollup/plugin-node-resolve": "^15.3.0",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/plugin-typescript": "^12.1.0",
+    "@types/gh-pages": "^6",
+    "@types/he": "^1",
+    "@types/node": "^22.7.4",
+    "@typescript-eslint/eslint-plugin": "^8.8.0",
+    "@typescript-eslint/parser": "^8.8.0",
+    "chokidar": "^4.0.1",
+    "chromedriver": "^129.0.2",
+    "copyfiles": "^2.4.1",
+    "cssnano": "^7.0.6",
+    "cypress": "^13.15.0",
+    "cypress-ct-lit": "^0.5.0",
+    "eslint": "^9.12.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "glob": "^11.0.0",
+    "globals": "^15.10.0",
+    "helmet": "^8.0.0",
+    "http-server": "^14.1.1",
+    "husky": "^9.1.6",
+    "lint-staged": "^15.2.10",
+    "postcss": "^8.2.15",
+    "prettier": "^3.3.3",
+    "rollup": "4.24.0",
     "rollup-plugin-cleanup": "^3.2.1",
-    "rollup-plugin-module-replacement": "^1.1.0",
-    "rollup-plugin-postcss": "^3.1.1",
-    "rollup-plugin-string-html": "^1.0.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.27.3",
-    "tachometer": "^0.5.5",
-    "tslib": "^2.0.3",
-    "typescript": "^4.5.5"
+    "rollup-plugin-dts": "^6.1.1",
+    "rollup-plugin-inline-postcss": "^3.0.1",
+    "rollup-plugin-minify-html-literals": "^1.2.6",
+    "rollup-plugin-postcss": "^4.0.2",
+    "sass": "^1.79.4",
+    "stream-browserify": "3",
+    "tachometer": "^0.7.1",
+    "tslib": "~2.7.0",
+    "typescript": "~5.5.0",
+    "typescript-eslint": "^8.8.1",
+    "vite": "^5.4.8",
+    "vite-plugin-alias": "^0.1.1",
+    "vite-plugin-html": "^3.2.2",
+    "vite-plugin-node-polyfills": "^0.22.0",
+    "vite-plugin-standard-css-modules": "^0.0.11",
+    "vite-plugin-static-copy": "^2.0.0",
+    "vite-plugin-tsconfig-paths": "^1.4.1",
+    "vite-resolve-tsconfig-paths": "^0.0.8",
+    "vite-tsconfig-paths": "^5.0.1"
   },
-  "browserslist": [
-    ">2%",
-    "Edge > 14",
-    "Firefox > 63",
-    "Chrome > 67"
-  ],
+  "browserslist": "> 0.5%, last 2 versions, not dead",
   "externals": [
     "./src/server/config.js"
   ],
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "lint-staged": {
-    "src/**/*.{ts}": [
-      "eslint --fix",
-      "prettier --write"
-    ]
-  }
+  "packageManager": "yarn@4.5.0"
 }
