@@ -1,6 +1,4 @@
-
 describe('TreeComponent Test', () => {
-
   beforeEach(() => {
     cy.visit('/test');
     cy.wait(100);
@@ -11,39 +9,80 @@ describe('TreeComponent Test', () => {
   });
 
   it('Displays text set by array', () => {
-    cy.get('app-testbed').shadow().find('x-tree')
-                         .shadow().find('x-node').first()
-                         .find('x-atom').shadow().contains('aaa');
+    cy.get('app-testbed')
+      .shadow()
+      .find('x-tree')
+      .shadow()
+      .find('x-node')
+      .first()
+      .shadow()
+      .find('x-atom')
+      .shadow()
+      .contains('aaa');
   });
 
   it('Displays text set by nested array', () => {
-    cy.get('app-testbed').shadow().find('x-tree')
-                         .shadow().find('x-node').eq(1)
-                         .find('x-atom').shadow().contains('fiz');
+    cy.get('app-testbed')
+      .shadow()
+      .find('x-tree')
+      .shadow()
+      .find('x-node')
+      .eq(1)
+      .shadow()
+      .find('x-atom')
+      .shadow()
+      .contains('fiz');
   });
 
   it('Displays text set by dot syntax', () => {
-    cy.get('app-testbed').shadow().find('x-tree')
-                         .shadow().find('x-node').eq(2)
-                         .find('x-atom').shadow().contains('bbb');
+    cy.get('app-testbed')
+      .shadow()
+      .find('x-tree')
+      .shadow()
+      .find('x-node')
+      .eq(2)
+      .shadow()
+      .find('x-atom')
+      .shadow()
+      .contains('bbb');
   });
 
   it('Displays text set by bracket notation', () => {
-    cy.get('app-testbed').shadow().find('x-tree')
-                         .shadow().find('x-node').eq(3)
-                         .find('x-atom').shadow().contains('fuz');
+    cy.get('app-testbed')
+      .shadow()
+      .find('x-tree')
+      .shadow()
+      .find('x-node')
+      .eq(3)
+      .shadow()
+      .find('x-atom')
+      .shadow()
+      .contains('fuz');
   });
 
   it('Displays text set by shallow property', () => {
-    cy.get('app-testbed').shadow().find('x-tree')
-                         .shadow().find('x-node').eq(4)
-                         .find('x-atom').shadow().contains('ddd');
+    cy.get('app-testbed')
+      .shadow()
+      .find('x-tree')
+      .shadow()
+      .find('x-node')
+      .eq(4)
+      .shadow()
+      .find('x-atom')
+      .shadow()
+      .contains('ddd');
   });
 
   it('Displays text set by string in setState()', () => {
-    cy.get('app-testbed').shadow().find('x-tree')
-                         .shadow().find('x-node').last()
-                         .find('x-atom').shadow().contains('deep');
+    cy.get('app-testbed')
+      .shadow()
+      .find('x-tree')
+      .shadow()
+      .find('x-node')
+      .last()
+      .shadow()
+      .find('x-atom')
+      .shadow()
+      .contains('deep');
   });
-
 });
