@@ -1,10 +1,9 @@
 const posthtml = require('posthtml');
 const inline = require('posthtml-inline');
-const minifier = require('htmlnano');
 const fs = require('fs');
 const path = require('path');
 
-const replace = str => {
+const replace = (str) => {
   return new Promise((res, rej) => {
     try {
       str = str.replace(`//# sourceMappingURL=web-animations.min.js.map`, '');
@@ -19,7 +18,7 @@ const replace = str => {
   let input = await fs.readFileSync('dist/client/index.html', 'utf8');
   input = input.replace('<head>', '<head><base href="/readymade">');
   const result = posthtml([
-    inline({ root: path.join(__dirname, 'dist', 'client') })
+    inline({ root: path.join(__dirname, 'dist', 'client') }),
   ]).process(input);
   const output = await result;
   const filtered = await replace(output.html);
