@@ -1,17 +1,10 @@
-import {
-  Component,
-  css,
-  CustomElement,
-  html,
-  State
-} from './../../../modules/core';
+import { Component, css, CustomElement, html, State } from '@readymade/core';
 
 declare let Prism: any;
 
-const env = process.env.NODE_ENV || 'development';
-
 export class CodeState {
-  public type: string;
+  public type: string = '';
+  public language: string = '';
 }
 
 @Component({
@@ -19,8 +12,8 @@ export class CodeState {
   style: css`
     :host {
       display: block;
-      padding-top: 1em;
-      padding-bottom: 1em;
+      padding: 1em;
+      background: var(--ready-color-container-bg);
     }
     code[class*='language-'],
     pre[class*='language-'] {
@@ -36,7 +29,7 @@ export class CodeState {
       word-wrap: normal;
       font-family: 'Source Code Pro', 'Courier New', monospace;
       font-size: 14px;
-      font-weight: 500;
+      font-weight: 400;
       color: #e0e2e4;
       text-shadow: none;
     }
@@ -196,13 +189,17 @@ export class CodeState {
     }
   `,
   template: html`
-    <pre class="language-{{type}}"><code class="language-{{type}}"></code></pre>
+    <pre class="{{language}}"><code class="{{language}}"></code></pre>
     <slot hidden></slot>
-  `
+  `,
 })
 class RCodeComponent extends CustomElement {
   constructor() {
     super();
+  }
+
+  connectedCallback() {
+    this.onSlotChange();
   }
 
   @State()
@@ -210,30 +207,27 @@ class RCodeComponent extends CustomElement {
     return new CodeState();
   }
 
-  public connectedCallback(event) {
-    this.onSlotChange();
-  }
-  
-  get observedAttributes() {
-    return ['type'];
+  static get observedAttributes() {
+    return ['language'];
   }
 
   public attributeChangedCallback(name, oldValue, newValue) {
     switch (name) {
-      case 'type':
-        this.setState('type', this.getAttribute('type'));
+      case 'language':
+        this.setState('type', newValue);
+        this.setState('language', `language-${newValue}`);
         break;
     }
   }
 
   public onSlotChange() {
-    const code = (this.shadowRoot.querySelector('slot').assignedNodes() as any)[
-      env === 'production' ? 0 : 1
-    ].textContent;
+    const code = (
+      this.shadowRoot.querySelector('slot').assignedNodes() as any
+    )[1].textContent;
     this.shadowRoot.querySelector('code').innerHTML = Prism.highlight(
       code,
       Prism.languages[this.getAttribute('type')],
-      this.getAttribute('type')
+      this.getAttribute('type'),
     );
   }
 }
