@@ -1,10 +1,14 @@
-import { Component, State, CustomElement } from './../../../modules/core';
+import { Component, State, CustomElement } from '@readymade/core';
+
+const CounterState = {
+  count: 0,
+};
 
 @Component({
   selector: 'my-counter',
   template: `
     <button id="dec">-</button>
-    <span>{{ c }}</span>
+    <span>{{ count }}</span>
     <button id="inc">+</button>
   `,
   style: `
@@ -27,12 +31,9 @@ import { Component, State, CustomElement } from './../../../modules/core';
 		background-color: seagreen;
 		color: white;
 	}
-	`
+	`,
 })
 export class MyCounter extends CustomElement {
-  private $state: {
-    c: '0';
-  };
   connectedCallback() {
     this.shadowRoot
       .querySelector('#inc')
@@ -42,20 +43,16 @@ export class MyCounter extends CustomElement {
       .addEventListener('click', this.dec.bind(this));
   }
 
-  get count() {
-    return Number.parseInt(this.$state.c);
+  @State()
+  public getState() {
+    return CounterState;
   }
 
   inc() {
-    this.setState('c', this.count + 1);
+    this.setState('count', this.getState().count + 1);
   }
 
   dec() {
-    this.setState('c', this.count - 1);
-  }
-
-  @State()
-  public getState() {
-    return { c: '0' };
+    this.setState('count', this.getState().count - 1);
   }
 }
