@@ -1,46 +1,49 @@
-import {
-  Component,
-  css,
-  CustomElement,
-  html,
-  State
-} from './../../../modules/core';
+import { Component, CustomElement, State } from '@readymade/core';
+
+const style = `
+  :host {
+    font-size: 16px;
+  }
+  h1 {
+    font-family: 'Major Mono Display', sans-serif;
+    padding-left: 1em;
+    -webkit-font-smoothing: antialiased;
+    -moz-osx-font-smoothing: grayscale;
+    -webkit-margin-before: 0px;
+    -webkit-margin-after: 0px;
+  }
+  h1,
+  span {
+    font-size: 1em;
+    letter-spacing: -0.04em;
+    margin-block-start: 0em;
+    margin-block-end: 0em;
+  }
+  h1.is--default,
+  span.is--default {
+    font-size: 1em;
+  }
+  h1.is--small,
+  span.is--small {
+    font-size: 12px;
+  }
+  h1.is--medium,
+  span.is--medium {
+    font-size: 6em;
+  }
+  h1.is--large,
+  span.is--large {
+    font-size: 12em;
+    padding-left: 0em;
+  }
+`;
+
+const template = `<h1 class="{{model.size}}">{{ model.copy }}</h1>`;
 
 @Component({
   selector: 'r-headline',
-  style: css`
-    h1 {
-      font-family: 'Major Mono Display', sans-serif;
-      padding-left: 1em;
-      -webkit-font-smoothing: antialiased;
-      -moz-osx-font-smoothing: grayscale;
-      -webkit-margin-before: 0px;
-      -webkit-margin-after: 0px;
-    }
-    h1,
-    span {
-      font-size: 1em;
-      letter-spacing: -0.04em;
-      margin-block-start: 0em;
-      margin-block-end: 0em;
-    }
-    h1.is--small,
-    span.is--small {
-      font-size: 12px;
-    }
-    h1.is--medium,
-    span.is--medium {
-      font-size: 6em;
-    }
-    h1.is--large,
-    span.is--large {
-      font-size: 12em;
-      padding-left: 0em;
-    }
-  `,
-  template: html`
-    <h1 class="{{model.size}}">{{ model.copy }}</h1>
-  `
+  style,
+  template,
 })
 class RHeadlineComponent extends CustomElement {
   public hyperNode: any;
@@ -58,8 +61,8 @@ class RHeadlineComponent extends CustomElement {
     return {
       model: {
         size: '',
-        copy: ''
-      }
+        copy: '',
+      },
     };
   }
 
@@ -78,4 +81,15 @@ class RHeadlineComponent extends CustomElement {
   }
 }
 
-export { RHeadlineComponent };
+const render = ({ size, copy }) => `
+  <r-headline>
+    <template shadowrootmode="open">
+      <style>
+      ${style}
+      </style>
+      <h1 class="${size}">${copy}</h1>
+    </template>
+  </r-headline>
+`;
+
+export { RHeadlineComponent, render };
