@@ -1,10 +1,4 @@
-import {
-  Component,
-  css,
-  CustomElement,
-  html,
-  State
-} from './../../../../modules/core';
+import { Component, css, CustomElement, html, State } from '@readymade/core';
 
 export class TreeState {
   public arrayModel = [
@@ -16,24 +10,24 @@ export class TreeState {
     'Node 5',
     'Node 6',
     'Node 7',
-    ['far', 'fiz', 'faz', 'fuz']
+    ['far', 'fiz', 'faz', 'fuz'],
   ];
   public objectModel = {
     foo: {
       bar: {
-        baz: 'bbb'
+        baz: 'bbb',
       },
       far: {
         fiz: {
           faz: {
-            fuz: 'fuz'
-          }
-        }
+            fuz: 'fuz',
+          },
+        },
       },
       mar: {
-        maz: 'mmm'
-      }
-    }
+        maz: 'mmm',
+      },
+    },
   };
   public ax = 'aaa';
   public bx = 'bbb';
@@ -49,8 +43,6 @@ export class TreeState {
     };
   };
 }
-
-export const _treeState = new TreeState();
 
 @Component({
   selector: 'x-tree',
@@ -72,7 +64,7 @@ export const _treeState = new TreeState();
     <x-node model="{{gx}}"></x-node>
     <x-node model="{{hx}}"></x-node>
     <x-node model="{{state.foo.bar}}"></x-node>
-  `
+  `,
 })
 class TreeComponent extends CustomElement {
   constructor() {
@@ -81,7 +73,49 @@ class TreeComponent extends CustomElement {
 
   @State()
   public getState() {
-    return new TreeState();
+    return {
+      arrayModel: [
+        'aaa',
+        'Node 1',
+        'Node 2',
+        'Node 3',
+        'Node 4',
+        'Node 5',
+        'Node 6',
+        'Node 7',
+        ['far', 'fiz', 'faz', 'fuz'],
+      ],
+      objectModel: {
+        foo: {
+          bar: {
+            baz: 'bbb',
+          },
+          far: {
+            fiz: {
+              faz: {
+                fuz: 'fuz',
+              },
+            },
+          },
+          mar: {
+            maz: 'mmm',
+          },
+        },
+      },
+      ax: 'aaa',
+      bx: 'bbb',
+      cx: 'ccc',
+      dx: 'ddd',
+      ex: 'eee',
+      fx: 'fff',
+      gx: 'ggg',
+      hx: 'hhh',
+      state: {
+        foo: {
+          bar: 'x',
+        },
+      },
+    };
   }
 
   static get observedAttributes() {
