@@ -1,5 +1,5 @@
 // test components
-export { ButtonState, MyButtonComponent } from './component/button';
+export { MyButtonComponent } from './component/button';
 export { RCodeComponent } from './component/code';
 export { MyCounter } from './component/counter';
 export { RHeadlineComponent } from './component/headline';
@@ -24,11 +24,14 @@ export {
   RdTextArea,
   RdSlider,
   RdSwitch,
-  RdDropdown
-} from './../../modules/ui';
+  RdDropdown,
+  RdButtonPad,
+} from '@readymade/ui';
 // views
 export { HomeComponent } from './view/home';
 export { PerformanceTestComponent } from './view/perf';
 export { QueryComponent } from './view/query';
 export { TestBedComponent } from './view/test';
 export { LibraryComponent } from './view/lib';
+export { NotFoundComponent } from './view/404';
+export { Router, routing } from './routing';
