@@ -1,13 +1,13 @@
-import { Router } from '../../modules/router';
+import { Router } from '@readymade/router';
 
 const routing = [
   { path: '/', component: 'app-home', title: 'Readymade' },
   { path: '/test', component: 'app-testbed', title: 'Readymade Test Page' },
-  { path: '/ui', component: 'app-library', title: 'Readymade UI' },
+  { path: '/lib', component: 'app-library', title: 'Readymade UI' },
   {
     path: '/perf',
     component: 'app-perftest',
-    title: 'Readymade Performance Test'
+    title: 'Readymade Performance Test',
   },
   {
     path: '/router',
@@ -15,10 +15,15 @@ const routing = [
     queryParams: {
       contentType: 'post',
       page: '1',
-      header: '1'
+      header: '1',
     },
-    title: 'Readymade Router Test'
-  }
+    title: 'Readymade Router Test',
+  },
+  {
+    path: '/404',
+    component: 'app-notfound',
+    title: 'File Not Found',
+  },
 ];
 
 export { Router, routing };
