@@ -1,1 +1,17 @@
-export { HomeComponent } from './home';
+export { ButtonState, MyButtonComponent } from './../../component/button';
+export { RCodeComponent } from './../../component/code';
+export { MyCounter } from './../../component/counter';
+export { RHeadlineComponent } from './../../component/headline';
+export { MyInputComponent } from './../../component/input';
+export { MyItemComponent } from './../../component/item';
+export { MyListComponent } from './../../component/list';
+export { RLogoComponent } from './../../component/logo';
+export { RMainNavComponent } from './../../component/main-nav';
+// docs
+export { RMeterComponent } from './../../component/meter';
+export { RSideNavComponent } from './../../component/side-nav';
+export { RStatsComponent } from './../../component/stats';
+export { AtomComponent } from './../../component/tree/atom';
+export { NodeComponent } from './../../component/tree/node';
+export { TreeComponent } from './../../component/tree/tree';
+export { HomeComponent, render } from './home';
