@@ -1,3 +1,1 @@
-export * from '@ungap/custom-elements';
-export * from 'element-internals-polyfill';
-export * from 'broadcastchannel-polyfill';
+import '@ungap/custom-elements';
