@@ -1,6 +1,11 @@
 export { attachDOM, attachShadow, attachStyle, define } from './src/attach';
 
-export { bindTemplate, compileTemplate, templateId, uuidv4 } from './src/compile';
+export {
+  bindTemplate,
+  compileTemplate,
+  templateId,
+  uuidv4,
+} from './src/compile';
 
 export {
   getSiblings,
@@ -8,5 +13,6 @@ export {
   getParent,
   querySelector,
   querySelectorAll,
-  getChildNodes
+  getChildNodes,
+  isNode,
 } from './src/util';
