@@ -1,6 +1,6 @@
 import { ElementMeta } from './../../decorator';
 
-function closestRoot(base: Element) {
+export function closestRoot(base: Element) {
   function __closestFrom(el: any): Element | HTMLHeadElement {
     if (el.getRootNode()) {
       return el.getRootNode();
@@ -11,23 +11,35 @@ function closestRoot(base: Element) {
   return __closestFrom(base);
 }
 
-function attachShadow(instance: any, options?: any) {
-  const shadowRoot: ShadowRoot = instance.attachShadow(options || {});
-  const t = document.createElement('template');
-  t.innerHTML = instance.template;
-  shadowRoot.appendChild(t.content.cloneNode(true));
+export function attachShadow(instance: any, options?: any) {
+  if (!instance.template) {
+    return;
+  }
+  if (!instance.shadowRoot) {
+    const shadowRoot: ShadowRoot = instance.attachShadow(options || {});
+    const t = document.createElement('template');
+    t.innerHTML = instance.template;
+    shadowRoot.appendChild(t.content.cloneNode(true));
+  }
   instance.bindTemplate();
 }
 
-function attachDOM(instance: any, options?: any) {
+export function attachDOM(instance: any) {
+  if (!instance.elementMeta) {
+    return;
+  }
   const t = document.createElement('template');
   t.innerHTML = instance.elementMeta.template;
   instance.appendChild(t.content.cloneNode(true));
   instance.bindTemplate();
 }
 
-function attachStyle(instance: any, options?: any) {
+export function attachStyle(instance: any) {
+  if (!instance.elementMeta) {
+    return;
+  }
   const id = `${instance.elementMeta.selector}`;
+
   const closest: any = closestRoot(instance);
   if (closest.tagName === 'HEAD' && document.getElementById(`${id}-x`)) {
     return;
@@ -42,7 +54,7 @@ function attachStyle(instance: any, options?: any) {
   closest.appendChild(t);
 }
 
-function define(instance: any, meta: ElementMeta) {
+export function define(instance: any, meta: ElementMeta) {
   if (meta.autoDefine === true) {
     if (meta.selector && !meta.custom) {
       customElements.define(meta.selector, instance.contructor);
@@ -53,5 +65,3 @@ function define(instance: any, meta: ElementMeta) {
     }
   }
 }
-
-export { attachDOM, attachStyle, attachShadow, define };
