@@ -1,12 +1,17 @@
 // events
+import { ElementMeta } from '../decorator';
 
-interface EmitterEvents {
+export interface EmitterEvents {
   [key: string]: any;
 }
 
-class ReadymadeEventTarget extends EventTarget {}
+export class ReadymadeEventTarget extends EventTarget {}
 
-class EventDispatcher {
+interface ReadymadeElementTarget extends Element {
+  elementMeta?: ElementMeta;
+}
+
+export class EventDispatcher {
   public target: Element;
   public events: {
     [key: string]: CustomEvent<any> | Event;
@@ -14,10 +19,11 @@ class EventDispatcher {
   public channels: {
     [key: string]: BroadcastChannel;
   };
+
   constructor(context: any, channelName?: string) {
     this.target = context;
     this.channels = {
-      default: new BroadcastChannel('default')
+      default: new BroadcastChannel('default'),
     };
     if (channelName) {
       this.setChannel(channelName);
@@ -27,6 +33,7 @@ class EventDispatcher {
   public get(eventName: string) {
     return this.events[eventName];
   }
+
   public set(eventName: string, ev: CustomEvent<any> | Event) {
     this.events[eventName] = ev;
     return this.get(eventName);
@@ -37,6 +44,7 @@ class EventDispatcher {
     }
     this.target.dispatchEvent(ev);
   }
+
   public broadcast(ev: CustomEvent<any> | Event | string, name?: string) {
     if (typeof ev === 'string') {
       ev = this.events[ev];
@@ -49,19 +57,22 @@ class EventDispatcher {
       defaultPrevented: ev.defaultPrevented,
       detail: (ev as CustomEvent).detail,
       timeStamp: ev.timeStamp,
-      type: ev.type
+      type: ev.type,
     };
-    name
-      ? this.channels[name].postMessage(evt)
-      : this.channels.default.postMessage(evt);
+    if (name) {
+      this.channels[name].postMessage(evt);
+    } else {
+      this.channels.default.postMessage(evt);
+    }
   }
   public setChannel(name: string) {
     this.channels[name] = new BroadcastChannel(name);
-    this.channels[name].onmessage = ev => {
-      for (const prop in (this.target as any).elementMeta.eventMap) {
+    this.channels[name].onmessage = (ev) => {
+      for (const prop in (this.target as ReadymadeElementTarget).elementMeta
+        ?.eventMap) {
         if (prop.includes(name) && prop.includes(ev.data.type)) {
-          this.target[(this.target as any).elementMeta.eventMap[prop].handler](
-            ev.data
+          this.target[(this.target as any).elementMeta?.eventMap[prop].handler](
+            ev.data,
           );
         }
       }
@@ -72,5 +83,3 @@ class EventDispatcher {
     delete this.channels[name];
   }
 }
-
-export { EventDispatcher, ReadymadeEventTarget };
