@@ -1,79 +1,32 @@
 import typescript from '@rollup/plugin-typescript';
 import resolve from '@rollup/plugin-node-resolve';
 import cleanup from 'rollup-plugin-cleanup';
-import { terser } from 'rollup-plugin-terser';
+import terser from '@rollup/plugin-terser';
 
 const clean = {
   comments: ['none'],
-  extensions: ['ts', 'js']
+  extensions: ['ts', 'js'],
 };
 
 export default [
   {
     input: 'src/modules/core/index.ts',
-    plugins: [resolve(), typescript({ declaration: false }), cleanup(clean)],
-    onwarn: (warning, next) => {
-      if (warning.code === 'THIS_IS_UNDEFINED') return;
-      next(warning);
-    },
-    output: {
-      file: 'packages/@readymade/core/fesm2015/core.js',
-      format: 'esm',
-      sourcemap: false
-    }
-  },
-  {
-    input: 'src/modules/core/index.ts',
     plugins: [
       resolve(),
       typescript({
-        sourceMap: true
+        sourceMap: false,
+        declarationDir: 'dist/packages/@readymade/core/fesm2022/typings',
       }),
       cleanup(clean),
-      terser()
     ],
     onwarn: (warning, next) => {
       if (warning.code === 'THIS_IS_UNDEFINED') return;
       next(warning);
     },
     output: {
-      file: 'packages/@readymade/core/fesm2015/core.min.js',
+      file: 'dist/packages/@readymade/core/fesm2022/index.js',
       format: 'esm',
-      sourcemap: true
-    }
-  },
-  {
-    input: 'src/modules/core/index.ts',
-    plugins: [resolve(), typescript({ declaration: false }), cleanup(clean)],
-    onwarn: (warning, next) => {
-      if (warning.code === 'THIS_IS_UNDEFINED') return;
-      next(warning);
+      sourcemap: true,
     },
-    output: {
-      file: 'packages/@readymade/core/bundles/core.js',
-      format: 'cjs',
-      sourcemap: false
-    }
   },
-  {
-    input: 'src/modules/core/index.ts',
-    plugins: [
-      resolve(),
-      typescript({
-        declaration: false,
-        sourceMap: true
-      }),
-      cleanup(clean),
-      terser()
-    ],
-    onwarn: (warning, next) => {
-      if (warning.code === 'THIS_IS_UNDEFINED') return;
-      next(warning);
-    },
-    output: {
-      file: 'packages/@readymade/core/bundles/core.min.js',
-      format: 'cjs',
-      sourcemap: true
-    }
-  }
 ];
