@@ -3,7 +3,7 @@ import {
   DOT_BRACKET_NOTATION_REGEX,
   findValueByString,
   isObject,
-  stripTemplateString
+  stripTemplateString,
 } from '../../core/element/src/compile';
 import { TemplateComponent } from '../custom';
 
@@ -18,7 +18,7 @@ function changeNode(protoNode: Element, key: string, value: any) {
     const template = path.substring(path.search(DOT_BRACKET_NOTATION_REGEX));
     node.textContent = protoNode.textContent.replace(
       protoNode.textContent,
-      isObject(value) ? findValueByString(value, template) : value
+      isObject(value) ? findValueByString(value, template) : value,
     );
   }
 
@@ -37,11 +37,11 @@ function changeNode(protoNode: Element, key: string, value: any) {
           }
           const path = stripTemplateString(attribute.nodeValue);
           const template = path.substring(
-            path.search(DOT_BRACKET_NOTATION_REGEX)
+            path.search(DOT_BRACKET_NOTATION_REGEX),
           );
           protoNode.setAttribute(
             attr,
-            isObject(value) ? findValueByString(value, template) : value
+            isObject(value) ? findValueByString(value, template) : value,
           );
           const remove = attribute.nodeName || attribute.name;
           node.removeAttribute(remove);
@@ -55,14 +55,14 @@ function changeNode(protoNode: Element, key: string, value: any) {
         }
         const path = stripTemplateString(attributeValue);
         const template = path.substring(
-          path.search(DOT_BRACKET_NOTATION_REGEX)
+          path.search(DOT_BRACKET_NOTATION_REGEX),
         );
         node.setAttribute(
           attr,
           attributeValue.replace(
             attributeValue,
-            isObject(value) ? findValueByString(value, template) : value
-          )
+            isObject(value) ? findValueByString(value, template) : value,
+          ),
         );
       }
       const check = attribute.nodeName || attribute.name;
@@ -74,11 +74,31 @@ function changeNode(protoNode: Element, key: string, value: any) {
   protoNode.parentNode.appendChild(node);
 }
 
+function isJSON(str) {
+  if (typeof str !== 'string') {
+    return false;
+  }
+  try {
+    JSON.parse(str);
+    return true;
+  } catch (e) {
+    console.error(e);
+    return false;
+  }
+}
+
+interface ReadymadeElement extends Element {
+  $id: string;
+  $key: string;
+
+  onChange(change: any): void;
+}
+
 function renderTemplate(
-  elem: any,
+  elem: ReadymadeElement,
   template: HTMLTemplateElement,
   items: string,
-  previousNode?: any
+  previousNode?: Element,
 ): void {
   if (!elem.parentNode) {
     return;
@@ -92,21 +112,27 @@ function renderTemplate(
 
   const clone = template.content.cloneNode(true);
   const protoNode = (clone as Element).querySelector(`[repeat="${bound[1]}"]`);
+
   let $elem: any = elem;
+
   let model: any;
 
   for (; $elem && $elem !== document; $elem = $elem.parentNode) {
-    if ($elem.host && $elem.host.$state && $elem.host.$state[bound[2]]) {
-      model = JSON.parse($elem.host.$state[bound[2]]);
+    if ($elem?.host && $elem?.host?.$state && $elem?.host?.$state[bound[2]]) {
+      model = isJSON($elem.host.$state[bound[2]])
+        ? JSON.parse($elem.host.$state[bound[2]])
+        : $elem.host.$state[bound[2]];
       elem.$key = bound[2];
       $elem.host.ɵɵstate.$changes.addEventListener(
         'change',
         (ev: CustomEvent) => {
           elem.onChange(ev.detail);
-        }
+        },
       );
-    } else if ($elem.$state && $elem.$state[bound[2]]) {
-      model = JSON.parse($elem.$state[bound[2]]);
+    } else if ($elem?.$state && $elem?.$state[bound[2]]) {
+      model = isJSON($elem.$state[bound[2]])
+        ? JSON.parse($elem.$state[bound[2]])
+        : $elem.$state[bound[2]];
       elem.$key = bound[2];
       $elem.ɵɵstate.$changes.addEventListener('change', (ev: CustomEvent) => {
         elem.onChange(ev.detail);
@@ -145,7 +171,7 @@ function renderTemplate(
 
 @Component({
   selector: 'r-repeat',
-  custom: { extends: 'template' }
+  custom: { extends: 'template' },
 })
 export class TemplateRepeater extends TemplateComponent {
   $id: string = templateId() + uuidv4().slice(0, 6);
@@ -175,7 +201,13 @@ export class TemplateRepeater extends TemplateComponent {
 
   public render(): void {
     const previousTarget = this.remove();
-    renderTemplate(this, this, this.getAttribute('items'), previousTarget);
+    if (this.getAttribute('force') === 'true') {
+      setTimeout(() =>
+        renderTemplate(this, this, this.getAttribute('items'), previousTarget),
+      );
+    } else {
+      renderTemplate(this, this, this.getAttribute('items'), previousTarget);
+    }
   }
 
   public onChange(change: any) {
@@ -186,7 +218,7 @@ export class TemplateRepeater extends TemplateComponent {
 }
 
 @Component({
-  selector: 'r-repeatr'
+  selector: 'r-repeatr',
 })
 export class Repeater extends PseudoElement {
   $id: string = templateId() + uuidv4().slice(0, 6);
