@@ -1,11 +1,4 @@
-import {
-  Component,
-  Emitter,
-  EventDispatcher,
-  FormElement,
-  html,
-  css
-} from './../../../core';
+import { Component, Listen, FormElement, html, css } from '@readymade/core';
 
 @Component({
   selector: 'rd-button',
@@ -18,16 +11,22 @@ import {
     :host button {
       width: 72px;
       height: 36px;
-      border: 2px solid var(--color-border);
-      background-color: var(--color-bg);
+      border: 2px solid var(--ready-color-border);
+      background-color: var(--ready-color-bg);
       border-radius: 14px;
-      color: var(--color-default);
+      color: var(--ready-color-default);
       cursor: pointer;
       display: flex;
       align-items: center;
       justify-content: center;
     }
-    :host button .icon {
+    :host button .label {
+      -webkit-user-select: none; /* Safari */
+      -moz-user-select: none; /* Firefox */
+      -ms-user-select: none; /* Internet Explorer/Edge */
+      user-select: none; /* Standard syntax */
+    }
+    :host button .icon:not(.is--empty) {
       display: block;
       width: 22px;
       height: 22px;
@@ -36,7 +35,7 @@ import {
       min-height: 18px;
       border-radius: 8px;
     }
-    :host button.is--small .icon {
+    :host button.is--small .icon:not(.is--empty) {
       display: inline-block;
       width: 12px;
       height: 12px;
@@ -45,7 +44,7 @@ import {
       min-height: 32px;
       border-radius: 14px;
     }
-    :host button.is--medium .icon {
+    :host button.is--medium .icon:not(.is--empty) {
       display: inline-block;
       width: 22px;
       height: 22px;
@@ -54,36 +53,38 @@ import {
       min-height: 44px;
       border-radius: 18px;
     }
-    :host button.is--large .icon {
+    :host button.is--large .icon:not(.is--empty) {
       display: inline-block;
       width: 32px;
       height: 32px;
     }
     :host button:hover {
-      background-color: var(--color-bg);
-      border: 2px solid var(--color-highlight);
+      background-color: var(--ready-color-bg);
+      border: 2px solid var(--ready-color-highlight);
     }
     :host button:focus {
       outline: 0px;
       outline-offset: 0px;
-      background-color: var(--color-bg);
-      border: 2px solid var(--color-highlight);
+      background-color: var(--ready-color-bg);
+      border: 2px solid var(--ready-color-highlight);
     }
-    :host button:active {
+    :host button:active,
+    :host button.active {
       outline: 0px;
       outline-offset: 0px;
-      background-color: var(--color-selected);
-      border: 2px solid var(--color-highlight);
+      background-color: var(--ready-color-selected);
+      border: 2px solid var(--ready-color-highlight);
     }
     :host button[disabled] {
-      opacity: var(--opacity-disabled);
-      background: var(--color-disabled);
+      opacity: var(--ready-opacity-disabled);
+      background: var(--ready-color-disabled);
       cursor: not-allowed;
     }
     :host button[disabled]:hover,
     :host button[disabled]:focus,
-    :host button[disabled]:active {
-      border: 2px solid var(--color-border);
+    :host button[disabled]:active,
+    :host button[disabled].active {
+      border: 2px solid var(--ready-color-border);
       outline: none;
       box-shadow: none;
     }
@@ -93,24 +94,41 @@ import {
       <span class="icon"><slot name="icon"></slot></span>
       <span class="label"><slot name="label"></slot></span>
     </button>
-  `
+  `,
 })
 class RdButton extends FormElement {
+  channel: BroadcastChannel;
+  _type: 'submit' | 'reset' | 'button' = 'button';
   constructor() {
     super();
+    this.type = 'button';
   }
 
   static get observedAttributes() {
-    return ['type'];
+    return ['type', 'label', 'width', 'height', 'channel'];
   }
 
   attributeChangedCallback(name: string, old: string, next: string) {
     switch (name) {
       case 'type':
-        this.type = next;
+        this._type = next as 'submit' | 'reset' | 'button';
+        this.type = next as 'submit' | 'reset' | 'button';
         break;
       case 'value':
         this.value = next;
+        break;
+      case 'label':
+        (this.shadowRoot.querySelector('.label') as HTMLSpanElement).innerText =
+          next;
+        break;
+      case 'width':
+        this.shadowRoot.querySelector('button').style.width = next;
+        break;
+      case 'height':
+        this.shadowRoot.querySelector('button').style.height = next;
+        break;
+      case 'channel':
+        this.setChannel(next);
         break;
     }
   }
@@ -119,21 +137,44 @@ class RdButton extends FormElement {
     this.$elem.disabled = disabled;
   }
 
-  @Emitter('change')
   connectedCallback() {
-    this.$elem.onchange = (ev: Event) => {
-      console.log(ev);
-      this.emitter.emit(
-        new CustomEvent('change', {
-          bubbles: true,
-          composed: true,
-          detail: 'composed'
-        })
-      );
-      if (this.onchange) {
-        this.onchange(ev);
+    this.shadowRoot.querySelectorAll('span').forEach((spanElem) => {
+      const slot = spanElem.querySelector('slot');
+      if (slot && slot.assignedNodes().length === 0) {
+        spanElem.classList.add('is--empty');
+      }
+    });
+    this.$elem.onclick = () => {
+      if (this.channel) {
+        this.channel.postMessage({
+          type: this.type,
+          name: this.name,
+          value: this.value.length ? this.value : 'bang',
+        });
       }
     };
+    if (this.type === 'submit') {
+      this.$elem.onsubmit = (ev: Event) => {
+        this.emitter.emit(
+          new CustomEvent('submit', {
+            bubbles: true,
+            composed: true,
+            detail: 'composed',
+          }),
+        );
+        if (this.onsubmit) {
+          this.onsubmit(ev as SubmitEvent);
+        }
+      };
+    }
+  }
+
+  @Listen('press')
+  onPress(ev: CustomEvent) {
+    if (ev.detail?.modifier) {
+      this.setAttribute('modifier', ev.detail?.modifier);
+    }
+    this.simulatePress();
   }
 
   get form() {
@@ -157,10 +198,10 @@ class RdButton extends FormElement {
   }
 
   get type() {
-    return this.$elem.type || 'button';
+    return this.$elem.type || this._type;
   }
 
-  set type(value) {
+  set type(value: 'submit' | 'reset' | 'button') {
     this.$elem.type = value;
   }
 
@@ -174,6 +215,19 @@ class RdButton extends FormElement {
 
   get $elem(): HTMLButtonElement {
     return this.shadowRoot.querySelector('button');
+  }
+
+  setChannel(name: string) {
+    this.channel = new BroadcastChannel(name);
+  }
+
+  simulatePress() {
+    this.$elem.classList.add('active');
+    this.$elem.click();
+    setTimeout(() => {
+      this.$elem.classList.remove('active');
+      this.removeAttribute('modifier');
+    }, 100);
   }
 }
 
