@@ -1,11 +1,4 @@
-import {
-  Component,
-  Emitter,
-  EventDispatcher,
-  FormElement,
-  html,
-  css
-} from './../../../core';
+import { Component, Emitter, FormElement, html, css } from '@readymade/core';
 
 @Component({
   selector: 'rd-checkbox',
@@ -28,12 +21,12 @@ import {
       display: block;
       width: 24px;
       height: 24px;
-      border: 2px solid var(--color-border);
+      border: 2px solid var(--ready-color-border);
       border-radius: 6px;
-      background: var(--color-bg);
+      background: var(--ready-color-bg);
     }
     :host input[type='checkbox']:checked:before {
-      background-image: var(--icon-check);
+      background-image: var(--ready-icon-check);
       background-repeat: no-repeat;
       background-position: center;
     }
@@ -45,22 +38,22 @@ import {
     :host input[type='checkbox']:hover:before,
     :host input[type='checkbox']:focus:before,
     :host input[type='checkbox']:active:before {
-      border: 2px solid var(--color-highlight);
+      border: 2px solid var(--ready-color-highlight);
     }
     :host input[type='checkbox'][disabled]:before {
-      opacity: var(--opacity-disabled);
-      background: var(--color-disabled);
+      opacity: var(--ready-opacity-disabled);
+      background: var(--ready-color-disabled);
       cursor: not-allowed;
     }
     :host input[type='checkbox'][disabled]:checked:before {
-      background-image: var(--icon-check);
+      background-image: var(--ready-icon-check);
       background-repeat: no-repeat;
       background-position: center;
     }
     :host input[type='checkbox'][disabled]:hover:before,
     :host input[type='checkbox'][disabled]:focus:before,
     :host input[type='checkbox'][disabled]:active:before {
-      border: 2px solid var(--color-border);
+      border: 2px solid var(--ready-color-border);
       outline: none;
       box-shadow: none;
     }
@@ -68,28 +61,30 @@ import {
     :host input[type='checkbox'].required:hover:before,
     :host input[type='checkbox'].required:focus:before,
     :host input[type='checkbox'].required:active:before {
-      border: 2px solid var(--color-error);
+      border: 2px solid var(--ready-color-error);
       outline: none;
       box-shadow: none;
     }
   `,
-  template: html`
-    <input type="checkbox" />
-  `
+  template: html` <input type="checkbox" /> `,
 })
 class RdCheckBox extends FormElement {
+  channel: BroadcastChannel;
   constructor() {
     super();
   }
 
   static get observedAttributes() {
-    return ['checked'];
+    return ['checked', 'channel'];
   }
 
   attributeChangedCallback(name: string, old: string, next: string) {
     switch (name) {
       case 'checked':
         this.checked = next === 'true' || next === '' ? true : false;
+        break;
+      case 'channel':
+        this.setChannel(next);
         break;
     }
   }
@@ -122,12 +117,19 @@ class RdCheckBox extends FormElement {
           new CustomEvent('change', {
             bubbles: true,
             composed: true,
-            detail: 'composed'
-          })
+            detail: 'composed',
+          }),
         );
       }
+      if (this.channel) {
+        this.channel.postMessage({
+          type: this.type,
+          name: this.name,
+          value: (ev.target as HTMLInputElement).checked,
+        });
+      }
     };
-    this.$elem.onblur = (ev: Event) => {
+    this.$elem.onblur = () => {
       this.onValidate();
     };
   }
@@ -180,6 +182,10 @@ class RdCheckBox extends FormElement {
 
   get $elem(): HTMLInputElement {
     return this.shadowRoot.querySelector('input');
+  }
+
+  setChannel(name: string) {
+    this.channel = new BroadcastChannel(name);
   }
 }
 
